Major Changes to Cards into Sheets
Many cards were removed and turned into sprite sheets. This was done to
fix issues with starting equipment being off sheets while cards in decks
were single sprites. This has been primarily fixed for all loose cards
in the game.

Game has been updated to deal with heroes gaining the target bounty.
Currently, the rewards are still showing the message for total damage,
etc. Will need to figure out how to remove that while the player is
picking a reward. Also will need to move the placement of rewards on the
page a bit.