Finished award section moved onto updating movement
Got the award section working fairly well. I still need to check and see
if the items are in the correct heroes bag by creating the equip phase
option and checking round two. I've also added in the room encounter
cards, giving them names and a timer per card. I'll need to eventually
implement the "Onset" card as the first card and check when heroes have
not met the rooms conditions. This will wait while I work on my next
biggest task: updating token movement.

Updating movement needs to happen per square as I'll need to check
certain conditions per room and other various reasons. Currently, heroes
move in a straight line and this will be problematic when I need to
start moving monster tokens. I'm thinking that a recursive call would be
best to setup a list of available moves to the selected empty space.
This could even possibly check for spaces that are supposed to be
unreachable with monsters blocking LoS, etc.