Major Update to Token Movement
Added a new font size: 6 for problems dealing with floor highlight and
what their center X/Y location was on the board, pixel wise. This is
being used to debug the new movement issues. The problem so far is that
after a token moves the 1st time, the location was off by 1 pixel versus
the board square location the token was standing on. This has been
roughly fixed but more debugging will need to happen.

New update to movement. Instead of the tokens moving in a straight line
to their target the tokens now move in a more board game like fashion.
The tokens will move as far as possible in a diagonal until either X/Y
difference becomes 0, then the token will move in a straight line to the
destination. With that, a recursive function was built to find the best
path. Currently, the function doesn't check for monsters blocking the
way or other obstacles (lava, pits, etc). More changes will come to that
when debugging is required.