Finished Token Movement
Did a great deal of work on token movement and it appears that hero
tokens are moving as they should be. The necromancer reanimate is also
moving correctly now as well. Speaking of, I still need to fix problems
with necromancer turn when no reanimate is active and the action order
of the reanimate.

Since I've finished movement (for now) I'm moving on to the equip phase
of the turn. I've currently set a message but need to deal with button
clicks and swapping the appropriate gear.