Equipment Swapping
Updated the awarding shop item section so that the kill count is cleared
and master kill count is reset back to 1.

Set up the equipment phase of the hero turn.  It checks over the bag to
see if there is anything in there. If there is, it presents the player
with the window to swap out gear. Review the code to see what's truly
going on but if weapons/trinkets are full, the player will need to
select where to put the stuff into. Currently, the issue that needs to
be addressed is when a bag item is swapped in the source rectangle is
messed up. Review the swapping code or the drawing code to find the
problem next time.